+++ docs/Changelog.md
@@ -2,11 +2,12 @@
 
 ## New since 0.6.0
 
-- POI API [#728](https://github.com/mapsforge/mapsforge/issues/728)
+- POI Search API [#728](https://github.com/mapsforge/mapsforge/issues/728)
 - Reverse Geocoding example [#727](https://github.com/mapsforge/mapsforge/issues/727)
 - LabelLayer enhanced implementation [#763](https://github.com/mapsforge/mapsforge/issues/763)
 - Color filters in map rendering [#778](https://github.com/mapsforge/mapsforge/issues/778)
 - Night mode example [#777](https://github.com/mapsforge/mapsforge/issues/777)
+- Java onTap layer listener [#774](https://github.com/mapsforge/mapsforge/issues/774)
 - Library integration [documentation](Integration.md) [#781](https://github.com/mapsforge/mapsforge/issues/781)
 - Jar with dependencies building [#767](https://github.com/mapsforge/mapsforge/issues/767)
 - Sample applications reorganization [#724](https://github.com/mapsforge/mapsforge/issues/724)